Fix potential UB in iterator
This commit avoids calling 'ptr::offset' with a size of '0'

'ptr::offset' requires (among other things) that the starting pointer be
in-bounds or one past the end of the allocation. If the pointer is null
or dangling (which can occur when we have a zero-sized allocation), then
it's impossible for either of those conditions to ever be met.

It's sufficient to check that 'inner_offset' is non-zero - if we have a non-zero inner_offset
with a dangling/null pointer, something has already gone wrong, and
there's nothing we can do about it.